renamed props,carousel and carouselbuttons,also renamed the customHeight to CarouselCardHeight and used tailwind css for the Image
