+++ app/page.tsx
@@ -9,6 +9,7 @@ const SLIDE_COUNT = 5
 const SLIDES = Array.from(Array(SLIDE_COUNT).keys())
 
 export default function Home() {
+
     return (
         <div>
             <Hero />
