+++ components/Carousel/SlideCard.tsx
@@ -12,12 +12,12 @@ const SlideCard = ({ index, name }: SlideCardProps) => {
             className={`rounded-3xl flex-0 flex-shrink-0 basis-[70%] max-w-sm lg:max-w-lg min-w-0 ${index > 0 ? '-ml-6' : '-ml-4'} relative`}
             style={{ zIndex: 5 + index }}
         >
-            <div className="relative border-orange border-4 rounded-3xl overflow-hidden sm:h-customHeight h-80">
+            <div className="relative border-orange border-4 rounded-3xl overflow-hidden sm:h-CarouselCardHeight h-80">
                 <Image
                     src={`/images/slide-${index + 1}.jpg`}
                     alt={`Slide ${index + 1}`}
                     fill={true}
-                    style={{ objectFit: "cover" }}
+                    className="object-cover"
                 />
             </div>
             <div className="absolute inset-0 bg-gradient-to-t from-black to-transparent opacity-50 rounded-3xl"></div>
