+++ package.json
@@ -13,6 +13,7 @@
     "embla-carousel-react": "^8.1.6",
     "next": "14.2.3",
     "react": "^18",
+    "react-calendar": "^5.0.0",
     "react-dom": "^18",
     "sharp": "^0.33.4"
   },
